requirements.txt
anthropic==0.40.0
flask==3.1.0
numpy==2.2.3
requests==2.32.3
httpx==0.28.1
psycopg2-binary==2.9.10
sqlalchemy==2.0.36
pgvector==0.3.5
redis==5.2.0
boto3==1.35.80
jinja2==3.1.4
python-dotenv==1.0.1
pydantic==2.10.3
tenacity==9.0.0
structlog==24.4.0
